+++ src/functions.lisp
@@ -3,7 +3,6 @@
 (defpackage cl-journal.functions
   (:use :cl :uiop/run-program)
   (:export :get-date-struct
-           :get-date-struct-str
            :cwd-is-git-dir-p
            :cwd-has-posts-file-p
            :exec
@@ -24,18 +23,6 @@
           :day date
           :hour hour
           :min minute)))
-
-(defun get-date-struct-str (&optional ts)
-  (multiple-value-bind
-        (second minute hour date month year)
-      (decode-universal-time (or ts (get-universal-time)))
-    (declare (ignore second))
-
-    (list "year" year
-          "mon" month
-          "day" date
-          "hour" hour
-          "min" minute)))
 
 (defun chdir (dir)
   (uiop/os::chdir dir)